Refactor hello redux store
