--- src/redux/action.js
@@ -1,4 +1,0 @@
-export const updateName = text => ({
-    type: 'USER_NAME_UPDATE',
-    text
-})
