+++ src/redux/hello/actions.js
@@ -1,0 +1,6 @@
+import { UPDATE_NAME_ACTION_TYPE } from './reducers';
+
+export const updateName = text => ({
+    type: UPDATE_NAME_ACTION_TYPE,
+    text
+});
